Part: Fix stack overflow in FCBRepAlgoAPI_BooleanOperation
In certain cases (boolean cut & one argument shape & one compound as tool shape) the Build() method causes an infinite
recursion and thus crashes because of a stack overflow. This is a regression of PR 19468.

Solution:
To avoid the recursion call the Build() method of the base class

The issue was reported at https://forum.freecad.org/viewtopic.php?t=95300