App: Fix possible crash in Document::removeObject()
Some important notes about transaction:
1. If within the same transaction an object is added and removed from a document then the call of removeObject() will destroy the document object.
2. If undo/redo is disabled then removeObject() will always destroy the document object.

The implementation of removeObject() first handles the transaction handling or destruction of the object and then it unregisters its label.
Therefore it uses a possibly dangling pointer to access the label. But this causes undefined behaviour and doesn't necessarily cause a direct crash.

Solution:
Before the transaction handling or destruction of the object store its label in a local string. Then use this string afterwards to unregister the
label.

This fixes issue 19980.