Sheet: Improve handling of unsupported formulas
With the example file of issue 20299 the parser stops with an exception and reads in the formulas only partially.
This PR doesn't fix the issue but avoids to raise an exception so that as much as possible will be read in.
For unsupported formulas a suitable message is printed with the content of the cell.