Material: Move definition of exception classes to source file
The separation of declaration and definition is needed to avoid implicit inlining as this can cause a crash on some platforms (e.g. OpenBSD).
The problem there is that for the shared libraries multiple symbols of the classes are generated so that the location where the exception is
thrown (in Materials.so) uses different symbols than in the location where it is supposed to be handled (in MatGui.so).

The system then claims that the thrown exception is nowhere handled and terminates the application.

This fixes issue 18170 and provides a better fix for issue 18168