BIM: Change Continue behavior and caching among commands
Currently, if we select `"Continue"` mode for BIM, it retriggers the
command, but it is not unique between commands and selecting it in one
place results in selecting it also in another place. Also, the place of
`Continue` is far below in the Tasks submenu.

So, this patch does a few things:
- removes additional "Continue" mode in BIM/Arch Mod, connecting to
  Draft's "Continue" mode directly (since we share those parts of GUI
  anyways)
- caches "Continue" mode separately between commands
- by doing the first point, this patch also moves "Continue" mode
  further above in the "Tasks" submenu