Gui: Improve handling of suffixes with FileDialog
The base class QFileDialog doesn't support complete suffixes but only simple suffixes and thus drops parts of a file extension.
E.g. if we have a file 'foobar.tar.gz' it drops the '.tar' so that it ends up with 'foobar.gz'.
This limitation will be fixed with this change.