Fix default argument in E57Exception
Assigning nullptr to a std::string is a bad idea because this usually causes a crash. Therefore in the constructor of
E57Exception the default argument shouldn't be nullptr but std::string() or "".
Luckily for some compilers the constructor std::string(nullptr_t) is 'deleted' and causes a build failure.

See forum: https://forum.freecad.org/viewtopic.php?t=99830