Part: Fix possible crash in section cut
createXBox() sets the value of the cutX field that may trigger the slot onCutXvalueChanged. This method calls deleteObejcts()
that deletes all created objects again. This makes the returned object by createBox() a dangling pointer and later may cause
a crash.

The same problem happens with the methods createYBox() and createZBox().

To fix the crash a QSignalBlocker is used to suppress the signaling