Brighter DefaultAmbientColor (#24059)
Currently when a user selects a custom Diffuse ShapeColor, the rendered
color is significantly too dim, which ends up being confusing and in
my view effectively being a usability issue.

The new DefaultAmbientColor (#d0d0d0) was determined by modelling an octagon
with both top and bottom chamfered, and angling the viewport on 45 degrees
on all three axis.

Then the Diffuse ShapeColor was changed so HSV(V=128), after which the
AmbientColor was raised until all 13 visible faces averaged HSV=(V=128)
on screen, resulting in the new DefaultAmbientColor (#d0d0d0)

To compensate for the new brighter AmbientColor, the DefaultShapeColor was
made more dim so it would roughly match HSV(V) as before.

Therefore this change should not be very noticable to most people, unless
they are changing their models Diffuse ShapeColor in which case they will
find a better matching render.