Core: Fix data races in redirector classes
When writing data to std::cout/std::cerr/std::clog and the RedirectStdOutput/RedirectStdError/RedirectStdLog classes are
used to redirect the output to the console class then data races can occur if the data are written in worker theads.

Example:
1. Load the FEMExample.FCStd file
2. Switch to the FEM workbench and purge the result objects
3. Wait for the auto-saving of the document
4. A heap-use-after-free error may occur
This is because several threads invoke PropertyPostDataObject::SaveDocFile() that prints an error message. Because the buffer
of RedirectStdError is not protected it can happen that overflow() is called in one thread while sync() is called in another
thread. So, the string passed to Console().Error() might be corrupted due to a re-allocation.

With a normal release or debug build usually no error occurs but if the address-sanitizer option is enabled it usually does.