Part: Handle exception of BRepLib::OrientClosedSolid
In TopoShape::fixSolidOrientation() the function BRepLib::OrientClosedSolid() is used to check and perform forward orientation
of the passed solid but may throw an exception with no description. This is fixed by setting a description to give the user
a feedback in which context it has failed.

See forum thread: https://forum.freecad.org/viewtopic.php?t=101181