Gui: Fix crash when using FreeCAD as Python module
When using FreeCADGui as module in Python then the order of deletion changes and the set of WorkbenchManipulator
is destroyed after the interpreter is finalized. Then to avoid a crash the GIL cannot be accessed any more.

To fix the problem a deleter is passed to the shared_ptr that doesn't delete the object if the interpreter is finalized.