Measure: Make measuring system more flexible
Currently the list of supported measure handlers of a module must be defined
inside the functions of MeasureClient. This is a violation of the open-closed
principal because any other module isn't able to extend this list so that the
file (in the Part module) must be adjusted over and over again.

To fix this limitation the method MeasureBaseExtendable::getMeasureInfo() is
changed to also search for sub-types of an object type.